emmm yea 5.5 bc nicknamez is gonna fuck me in ass and make fucking copy :skull:

+++ Updater/EzSploit Launcher/Form1.cs
@@ -22,6 +22,19 @@ namespace EzSploit_Launcher
         string selectedapi;
         string selectedtheme;
         string autoinj;
+        string script2;
+        string script3;
+        string script4;
+        string script5;
+        string script6;
+        string script7;
+        string script1text;
+        string script2text;
+        string script3text;
+        string script4text;
+        string script5text;
+        string script6text;
+        string script7text;
 
         WebClient webClient = new WebClient();
         public void wait(int milliseconds)
@@ -75,24 +88,69 @@ namespace EzSploit_Launcher
             wait(500);
             System.IO.File.Delete(@"c:\mikusdevPrograms\ezsploit\version.txt");
             wait(500);
-            System.IO.File.Move(@"c:\mikusdevPrograms\ezsploit\versionew.txt", @"c:\mikusdevPrograms\ezsploit\version.txt");
-            wait(500);
+            if (System.IO.File.Exists(@"c:\mikusdevPrograms\ezsploit\versionew.txt"))
+            {
+                System.IO.File.Move(@"c:\mikusdevPrograms\ezsploit\versionew.txt", @"c:\mikusdevPrograms\ezsploit\version.txt");
+                wait(500);
+            }
+            
             DirectoryInfo di1 = Directory.CreateDirectory(@"c:\mikusdevPrograms\ezsploit\updatetemp");
 
             textboxtext = System.IO.File.ReadAllText(@"c:\mikusdevPrograms\ezsploit\Configs\textboxconf.txt");
             selectedapi = System.IO.File.ReadAllText(@"c:\mikusdevPrograms\ezsploit\Configs\selectedAPI.txt");
             selectedtheme = System.IO.File.ReadAllText(@"c:\mikusdevPrograms\ezsploit\Configs\selectedTheme.txt");
             autoinj = System.IO.File.ReadAllText(@"c:\mikusdevPrograms\ezsploit\Configs\autoinject.txt");
+            script2 = System.IO.File.ReadAllText(@"c:\mikusdevPrograms\ezsploit\Configs\script2.txt");
+            script3 = System.IO.File.ReadAllText(@"c:\mikusdevPrograms\ezsploit\Configs\script3.txt");
+            script4 = System.IO.File.ReadAllText(@"c:\mikusdevPrograms\ezsploit\Configs\script4.txt");
+            script5 = System.IO.File.ReadAllText(@"c:\mikusdevPrograms\ezsploit\Configs\script5.txt");
+            script6 = System.IO.File.ReadAllText(@"c:\mikusdevPrograms\ezsploit\Configs\script6.txt");
+            script7 = System.IO.File.ReadAllText(@"c:\mikusdevPrograms\ezsploit\Configs\script7.txt");
+            script1text = System.IO.File.ReadAllText(@"c:\mikusdevPrograms\ezsploit\Configs\script1text.txt");
+            script2text = System.IO.File.ReadAllText(@"c:\mikusdevPrograms\ezsploit\Configs\script2text.txt");
+            script3text = System.IO.File.ReadAllText(@"c:\mikusdevPrograms\ezsploit\Configs\script3text.txt");
+            script4text = System.IO.File.ReadAllText(@"c:\mikusdevPrograms\ezsploit\Configs\script4text.txt");
+            script5text = System.IO.File.ReadAllText(@"c:\mikusdevPrograms\ezsploit\Configs\script5text.txt");
+            script6text = System.IO.File.ReadAllText(@"c:\mikusdevPrograms\ezsploit\Configs\script6text.txt");
+            script7text = System.IO.File.ReadAllText(@"c:\mikusdevPrograms\ezsploit\Configs\script7text.txt");
 
             using (FileStream fs = System.IO.File.Create(@"c:\mikusdevPrograms\ezsploit\updatetemp\textboxconf.txt"))
             using (FileStream fs2 = System.IO.File.Create(@"c:\mikusdevPrograms\ezsploit\updatetemp\selectedAPI.txt"))
             using (FileStream fs3 = System.IO.File.Create(@"c:\mikusdevPrograms\ezsploit\updatetemp\selectedTheme.txt"))
             using (FileStream fs4 = System.IO.File.Create(@"c:\mikusdevPrograms\ezsploit\updatetemp\autoinject.txt"))
+
+            using (FileStream fs23 = System.IO.File.Create(@"c:\mikusdevPrograms\ezsploit\updatetemp\script1.txt"))
+            using (FileStream fs22 = System.IO.File.Create(@"c:\mikusdevPrograms\ezsploit\updatetemp\script2.txt"))
+            using (FileStream fs32 = System.IO.File.Create(@"c:\mikusdevPrograms\ezsploit\updatetemp\script3.txt"))
+            using (FileStream fs42 = System.IO.File.Create(@"c:\mikusdevPrograms\ezsploit\updatetemp\script4.txt"))
+            using (FileStream fs221 = System.IO.File.Create(@"c:\mikusdevPrograms\ezsploit\updatetemp\script5.txt"))
+            using (FileStream fs321 = System.IO.File.Create(@"c:\mikusdevPrograms\ezsploit\updatetemp\script6.txt"))
+            using (FileStream fs421 = System.IO.File.Create(@"c:\mikusdevPrograms\ezsploit\updatetemp\script7.txt"))
+            using (FileStream fs233 = System.IO.File.Create(@"c:\mikusdevPrograms\ezsploit\updatetemp\script1text.txt"))
+            using (FileStream fs223 = System.IO.File.Create(@"c:\mikusdevPrograms\ezsploit\updatetemp\script2text.txt"))
+            using (FileStream fs323 = System.IO.File.Create(@"c:\mikusdevPrograms\ezsploit\updatetemp\script3text.txt"))
+            using (FileStream fs432 = System.IO.File.Create(@"c:\mikusdevPrograms\ezsploit\updatetemp\script4text.txt"))
+            using (FileStream fs2231 = System.IO.File.Create(@"c:\mikusdevPrograms\ezsploit\updatetemp\script5text.txt"))
+            using (FileStream fs3321 = System.IO.File.Create(@"c:\mikusdevPrograms\ezsploit\updatetemp\script6text.txt"))
+            using (FileStream fs4321 = System.IO.File.Create(@"c:\mikusdevPrograms\ezsploit\updatetemp\script7text.txt"))
             wait(100);
             System.IO.File.WriteAllText(@"c:\mikusdevPrograms\ezsploit\updatetemp\textboxconf.txt", textboxtext);
             System.IO.File.WriteAllText(@"c:\mikusdevPrograms\ezsploit\updatetemp\selectedAPI.txt", selectedapi);
             System.IO.File.WriteAllText(@"c:\mikusdevPrograms\ezsploit\updatetemp\selectedTheme.txt", selectedtheme);
             System.IO.File.WriteAllText(@"c:\mikusdevPrograms\ezsploit\updatetemp\autoinject.txt", autoinj);
+            System.IO.File.WriteAllText(@"c:\mikusdevPrograms\ezsploit\updatetemp\script2.txt", script2);
+            System.IO.File.WriteAllText(@"c:\mikusdevPrograms\ezsploit\updatetemp\script3.txt", script3);
+            System.IO.File.WriteAllText(@"c:\mikusdevPrograms\ezsploit\updatetemp\script4.txt", script4);
+            System.IO.File.WriteAllText(@"c:\mikusdevPrograms\ezsploit\updatetemp\script5.txt", script5);
+            System.IO.File.WriteAllText(@"c:\mikusdevPrograms\ezsploit\updatetemp\script6.txt", script6);
+            System.IO.File.WriteAllText(@"c:\mikusdevPrograms\ezsploit\updatetemp\script7.txt", script7);
+            System.IO.File.WriteAllText(@"c:\mikusdevPrograms\ezsploit\updatetemp\script1text.txt", script1text);
+            System.IO.File.WriteAllText(@"c:\mikusdevPrograms\ezsploit\updatetemp\script2text.txt", script2text);
+            System.IO.File.WriteAllText(@"c:\mikusdevPrograms\ezsploit\updatetemp\script3text.txt", script3text);
+            System.IO.File.WriteAllText(@"c:\mikusdevPrograms\ezsploit\updatetemp\script4text.txt", script4text);
+            System.IO.File.WriteAllText(@"c:\mikusdevPrograms\ezsploit\updatetemp\script5text.txt", script5text);
+            System.IO.File.WriteAllText(@"c:\mikusdevPrograms\ezsploit\updatetemp\script6text.txt", script6text);
+            System.IO.File.WriteAllText(@"c:\mikusdevPrograms\ezsploit\updatetemp\script7text.txt", script7text);
 
 
             Directory.Delete(@"c:\mikusdevPrograms\ezsploit\Configs", true);
